Work on new unit tests
- Rename unit tests files for clarity.
- Remove unused usings.
- Rename locals to follow project code style.
- Migrate using to modern pattern.
- Replace try/catch with Assert for object disposed exception.
- Minor code style fixes.
